[Slider]: add `triggerOnDragging` property for determine when to trigger `onChange` event on component.
In the old code, `onChange` event will be triggered on every mouse move. But this behavior will not be as expected in some scenario. So I added a property (`triggerOnDragging`) to control the it in different mode. By default it will be `false` as the `onChange` event will only be trigger when mouse is released. When set it to `true`, every movement of dragging mouse will trigger `onChange`.